+++ lib/datadog/appsec/waf/context.rb
@@ -74,7 +74,7 @@ module Datadog
 
           code = LibDDWAF.ddwaf_run(@context_ptr, persistent_data_obj, ephemeral_data_obj, result_obj, timeout)
 
-          Result.new(
+          result = Result.new(
             RESULT_CODE[code],
             Converter.object_to_ruby(result_obj[:events]),
             result_obj[:total_runtime],
@@ -82,6 +82,10 @@ module Datadog
             Converter.object_to_ruby(result_obj[:actions]),
             Converter.object_to_ruby(result_obj[:derivatives])
           )
+
+          result.mark_as_input_truncated! if persistent_data_obj.input_truncated? || ephemeral_data_obj.input_truncated?
+
+          result
         ensure
           LibDDWAF.ddwaf_result_free(result_obj) if result_obj
           LibDDWAF.ddwaf_object_free(ephemeral_data_obj) if ephemeral_data_obj