+++ lib/datadog/appsec/waf/result.rb
@@ -15,6 +15,16 @@ module Datadog
           @timeout = timeout
           @actions = actions
           @derivatives = derivatives
+
+          @input_truncated = false
+        end
+
+        def mark_as_input_truncated!
+          @input_truncated = true
+        end
+
+        def input_truncated?
+          @input_truncated
         end
 
         def to_h